techDebt(removeNewFromWorkplaceNotFoundComponents): Remove new from workplace-not-found components
