+++ src/app/features/add-workplace/add-workplace.module.ts
@@ -15,13 +15,13 @@ import { FindWorkplaceAddressComponent } from './find-workplace-address/find-wor
 import { FindYourWorkplaceComponent } from './find-your-workplace/find-your-workplace.component';
 import { IsThisYourWorkplaceComponent } from './is-this-your-workplace/is-this-your-workplace.component';
 import { NameOfWorkplaceComponent } from './name-of-workplace/name-of-workplace.component';
-import { NewWorkplaceNotFoundComponent } from './new-workplace-not-found/new-workplace-not-found.component';
 import { RegulatedByCqcComponent } from './regulated-by-cqc/regulated-by-cqc.component';
 import { SelectMainServiceComponent } from './select-main-service/select-main-service.component';
 import { SelectWorkplaceAddressComponent } from './select-workplace-address/select-workplace-address.component';
 import { SelectWorkplaceComponent } from './select-workplace/select-workplace.component';
 import { WorkplaceAddedThankYouComponent } from './workplace-added-thank-you/workplace-added-thank-you.component';
 import { WorkplaceNameAddressComponent } from './workplace-name-address/workplace-name-address.component';
+import { WorkplaceNotFoundComponent } from './workplace-not-found/workplace-not-found.component';
 
 @NgModule({
   imports: [CommonModule, AddWorkplaceRoutingModule, ReactiveFormsModule, SharedModule],
@@ -38,7 +38,7 @@ import { WorkplaceNameAddressComponent } from './workplace-name-address/workplac
     StartComponent,
     WorkplaceNameAddressComponent,
     NameOfWorkplaceComponent,
-    NewWorkplaceNotFoundComponent,
+    WorkplaceNotFoundComponent,
     IsThisYourWorkplaceComponent,
     FindYourWorkplaceComponent,
     SelectMainServiceComponent,
