+++ src/app/features/add-workplace/workplace-not-found/workplace-not-found.component.ts
@@ -4,24 +4,24 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { BackService } from '@core/services/back.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
 import { EstablishmentService } from '@core/services/establishment.service';
-import { RegistrationService } from '@core/services/registration.service';
+import { WorkplaceService } from '@core/services/workplace.service';
 import { NewWorkplaceNotFoundDirective } from '@shared/directives/create-workplace/new-workplace-not-found/new-workplace-not-found.directive';
 
 @Component({
-  selector: 'app-new-workplace-not-found',
+  selector: 'app-workplace-not-found',
   templateUrl:
-    '../../../../shared/directives/create-workplace/new-workplace-not-found/new-workplace-not-found.component.html',
+    '../../../shared/directives/create-workplace/new-workplace-not-found/new-workplace-not-found.component.html',
 })
-export class NewWorkplaceNotFoundComponent extends NewWorkplaceNotFoundDirective {
+export class WorkplaceNotFoundComponent extends NewWorkplaceNotFoundDirective {
   constructor(
     protected establishmentService: EstablishmentService,
     protected formBuilder: FormBuilder,
     public backService: BackService,
     protected errorSummaryService: ErrorSummaryService,
-    protected registrationService: RegistrationService,
+    protected workplaceService: WorkplaceService,
     protected router: Router,
     protected route: ActivatedRoute,
   ) {
-    super(establishmentService, formBuilder, backService, errorSummaryService, registrationService, router, route);
+    super(establishmentService, formBuilder, backService, errorSummaryService, workplaceService, router, route);
   }
 }
