+++ src/app/features/create-account/workplace/workplace-name-address/workplace-name-address.component.spec.ts
@@ -371,7 +371,7 @@ describe('WorkplaceNameAddressComponent', () => {
         component.setBackLink();
 
         expect(backLinkSpy).toHaveBeenCalledWith({
-          url: ['/registration', 'new-workplace-not-found'],
+          url: ['/registration', 'workplace-not-found'],
         });
       });
     });
