+++ src/app/features/registration/registration-routing.module.ts
@@ -21,8 +21,8 @@ import { RegistrationCompleteComponent } from '@features/registration/registrati
 
 import { FindYourWorkplaceComponent } from '../create-account/workplace/find-your-workplace/find-your-workplace.component';
 import { IsThisYourWorkplaceComponent } from '../create-account/workplace/is-this-your-workplace/is-this-your-workplace.component';
-import { NewWorkplaceNotFoundComponent } from '../create-account/workplace/new-workplace-not-found/new-workplace-not-found.component';
 import { RegulatedByCqcComponent } from '../create-account/workplace/regulated-by-cqc/regulated-by-cqc.component';
+import { WorkplaceNotFoundComponent } from '../create-account/workplace/workplace-not-found/workplace-not-found.component';
 import { AboutUsRegistrationComponent } from './about-us/about-us.component';
 import { CreateAccountComponent } from './create-account/create-account.component';
 
@@ -61,8 +61,8 @@ const routes: Routes = [
     data: { title: 'Service regulated by CQC?' },
   },
   {
-    path: 'new-workplace-not-found',
-    component: NewWorkplaceNotFoundComponent,
+    path: 'workplace-not-found',
+    component: WorkplaceNotFoundComponent,
     canActivate: [RegisterGuard],
     data: { title: 'Could not find your workplace' },
   },
